weefee/test_main.py
import unittest
from unittest.mock import patch
import getpass
from weefee.main import generate_weefee_qr,main

class TestGenerateWeefeeQR(unittest.TestCase):

    @patch('main.generate_weefee_qr')
    def test_generate_weefee_qr(self, mock_generate_weefee_qr):
        mock_generate_weefee_qr.return_value = None
        # Testing with WPA authentication, visible SSID, and a password
        # We expect the function to generate a QR code and print a message
        self.assertEqual(generate_weefee_qr("WPA", "MyWiFi", "password123", False), None)

        # Testing with no authentication, hidden SSID, and no password
        # We expect the function to generate a QR code and print a message
        self.assertEqual(generate_weefee_qr("nopass", "HiddenWiFi", None, True), None)

        # Testing with WEP authentication, visible SSID, and a password
        # We expect the function to generate a QR code and print a message
        self.assertEqual(generate_weefee_qr("WEP", "MyWiFi", "password123", False), None)


    @patch('main.getpass.getpass')
    def test_generate_weefee_qr2(self, mock_getpass):
        mock_getpass.side_effect = ['test_ssid', 'test_password']
        generate_weefee_qr("WPA", "test_ssid", "test_password", False)
        # Test case 1: Test with valid inputs (WPA, non-hidden SSID, valid SSID and password)
        
        mock_getpass.side_effect = ['test_ssid', '']
        generate_weefee_qr("WPA", "test_ssid", "", False)
        # Test case 2: Test with empty password
        
        mock_getpass.side_effect = ['', 'test_password']
        generate_weefee_qr("WPA", "", "test_password", False)
        # Test case 3: Test with empty SSID
        
        mock_getpass.side_effect = ['', '']
        generate_weefee_qr("WPA", "", "", False)
        # Test case 4: Test with empty SSID and password
        
        mock_getpass.side_effect = ['', '']
        generate_weefee_qr("WPA", "", "", True)
        # Test case 5: Test with empty SSID and password and hidden SSID

if __name__ == '__main__':
    unittest.main()